Encapsulate item damage via acessors

+++ src/crafting.cpp
@@ -1657,7 +1657,7 @@ void player::complete_disassemble( int item_pos, const tripoint &loc,
     // has been removed.
     item dis_item = org_item;
 
-    float component_success_chance = ( float )( std::min( std::pow( 0.8, dis_item.damage ), 1.0 ) );
+    float component_success_chance = std::min( std::pow( 0.8, dis_item.damage() ), 1.0 );
 
     int veh_part = -1;
     vehicle *veh = g->m.veh_at( pos(), veh_part );
